Move kcsession.py to be part of rest package
